PUBLIC: [p4-constraints] Add an e2e test for action constraints.
I refactored and extended interpreter_golden_test_runner for golden testing action constraints. I tested entries meeting and violating actions and action profile action sets. I also added tests for fields in the entry that are not in the constraint for an additional check that only relevant test output is printed.

PiperOrigin-RevId: 583206491